split resolve-lock
