+++ testcase/resolve-lock/resolve_lock.go
